separate Runner component into smaller components
