+++ src/components/store.js
@@ -1,2 +1,4 @@
 import { atom } from "nanostores";
+
 export const currentHeading = atom("");
+export const currentDirectory = atom("");
