+++ src/pages/run.astro
@@ -2,7 +2,7 @@
 import PageLayout from "@layouts/PageLayout.astro";
 import CardThin from "@components/CardThin.astro";
 import Spinner from "@components/Spinner.astro";
-import Runner from "@components/Runner.svelte";
+import Runner from "@components/run/Runner.svelte";
 import { Icon } from "astro-icon";
 ---
